feat: add image rule for both plugins
