+++ plugins/utils/redirect-resource.js
